+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format of the file is based on [Keep a Changelog](http://keepachangelog.com/
 - "What is/was new in C# X.Y?" link to MSDN documentation on C# version in the Sharpen Results view.
 - "Learn more..." link to MSDN documentation on language features in the Sharpen Results view.
 
+### Changed
+- "Async streams" renamed to "Asynchronous streams".
+
 ## [0.9.0] - 2019-06-14
 ### Added
 - "Consider replacing using statement with using declaration" suggestion.